# Conflicts:
#	resources/js/Components/ChannelSelectBar.vue
#	resources/js/Components/ServerSelectBar.vue
#	resources/js/Components/TextSelectBar.vue
#	resources/js/Layouts/AuthenticatedLayout.vue
#	resources/js/Pages/Text/Texting.vue
#	resources/js/Pages/Voice/Speaking.vue
#	resources/views/app.blade.php

+++ app/Http/Controllers/Api/ServerController.php
@@ -2,16 +2,18 @@
 
 namespace App\Http\Controllers\Api;
 
+use App\Enums\PermsType;
 use App\Events\Servers\ServerCreated;
 use App\Events\Servers\ServerEdited;
 use App\Events\Servers\ServerJoined;
 use App\Events\Servers\ServerLeft;
 use App\Http\Controllers\Controller;
+use App\Models\Role;
 use App\Models\Server;
 use Auth;
 use Illuminate\Http\Request;
-use Storage;
 use Inertia\Inertia;
+use Storage;
 
 class ServerController extends Controller
 {
@@ -20,7 +22,7 @@ class ServerController extends Controller
         $request->validate([
             'name' => 'required|string|max:50',
             'description' => 'nullable|string|max:500',
-            'icon' => 'nullable|image|mimes:jpeg,png,jpg|max:2048',
+            'icon' => 'nullable|image|mimes:jpeg,png,jpg,gif,webp|max:2048',
         ]);
 
         if ($request->file('icon')?->isValid()) {
@@ -33,9 +35,18 @@ class ServerController extends Controller
             'icon' => empty($path) ? null : Storage::url($path),
         ]);
 
-        $server->users()->attach(Auth::id());
+        $role = Role::create([
+            'name' => 'Owner',
+            'color' => '#ffffff',
+            'perms' => PHP_INT_MAX,
+            'importance' => 0,
+        ]);
 
-        broadcast(new ServerCreated($server->id, $server->name, $server->description, $server->icon));
+        $server->roles()->attach($role);
+        $server->users()->attach(Auth::id());
+        Auth::user()->roles()->attach($role);
+
+        //        broadcast(new ServerCreated($server->id, $server->name, $server->description, $server->icon));
 
         return response()->json(['message' => 'Server created successfully.']);
     }
@@ -49,11 +60,17 @@ class ServerController extends Controller
         $data = explode('#', $request->code);
         $serverId = $data[0];
         $code = $data[1];
-        if (hash('xxh32', $serverId) !== $code) return response()->json(['message' => 'Code is invalid.'], 404);
+        if (hash('xxh32', $serverId) !== $code) {
+            return response()->json(['message' => 'Code is invalid.'], 404);
+        }
 
         $server = Server::find($serverId);
-        if (!$server) return response()->json(['message' => 'Server not found.'], 404);
-        if ($server->users->has(Auth::id())) return response()->json(['message' => 'Server has already been added.'], 409);
+        if (! $server) {
+            return response()->json(['message' => 'Server not found.'], 404);
+        }
+        if ($server->users->has(Auth::id())) {
+            return response()->json(['message' => 'Server has already been added.'], 409);
+        }
 
         $server->users()->attach(Auth::id());
 
@@ -61,7 +78,6 @@ class ServerController extends Controller
 
         return response()->json(['message' => 'User added to server successfully.']);
     }
-
 
     public function removeUser(Request $request, int $serverId)
     {
@@ -71,8 +87,16 @@ class ServerController extends Controller
 
         $server = Server::find($serverId);
 
-        if (!$server) {
+        if (! $server) {
             return response()->json(['message' => 'Server not found.'], 404);
+        }
+
+        $roles = $server->roles->intersect(Auth::user()->roles);
+
+        if ($roles->doesntContain(function (Role $role) {
+            return $role->hasPerms(PermsType::CAN_KICK->value);
+        })) {
+            return response()->json(['message' => 'Forbidden.'], 403);
         }
 
         $server->users()->detach($request->user_id);
@@ -85,57 +109,24 @@ class ServerController extends Controller
     public function edit(Request $request, int $serverId)
     {
         $request->validate([
-            'name' => 'nullable|string|max:50',
+            'name' => 'required|string|max:50',
             'description' => 'nullable|string|max:500',
-            'icon' => 'nullable|string|max:255',
-        ]);
-
-        $server = Server::find($serverId);
-
-        if (!$server) {
-            return response()->json(['message' => 'Server not found.'], 404);
-        }
-
-        $server->update($request->only(['name', 'description', 'icon']));
-
-        broadcast(new ServerEdited($serverId, $server->name, $server->description, $server->icon));
-
-        return response()->json(['message' => 'Server updated successfully.']);
-    }
-
-    public function showSettings($serverId)
-    {
-        $server = Server::find($serverId);
-        if (!$server) {
-            return response()->json(['message' => 'Server not found.'], 404);
-        }
-
-        return Inertia::render('Settings/Server')->with(['selected_server' => $server]);
-    }
-
-    public function destroy(int $serverId)
-    {
-        $server = Server::find($serverId);
-
-        if (!$server) {
-            return response()->json(['message' => 'Server not found.'], 404);
-        }
-
-        $server->delete();
-
-        return redirect('/home');
-    }
-
-    public function update(Request $request, int $serverId)
-    {
-        $request->validate([
-            'name' => 'nullable|string|max:50',
-            'description' => 'nullable|string|max:255',
             'icon' => 'nullable|image|mimes:jpeg,png,jpg|max:2048',
         ]);
 
         $server = Server::find($serverId);
-        if (!$server) return redirect()->back()->withErrors(['message' => 'Server not found.']);
+
+        if (! $server) {
+            return response()->json(['message' => 'Server not found.'], 404);
+        }
+
+        $roles = $server->roles->intersect(Auth::user()->roles);
+
+        if ($roles->doesntContain(function (Role $role) {
+            return $role->hasPerms(PermsType::CAN_EDIT_SERVER->value);
+        })) {
+            return response()->json(['message' => 'Forbidden.'], 403);
+        }
 
         if ($request->file('icon')?->isValid()) {
             $path = $request->file('icon')->store('uploads', 'public');
@@ -146,7 +137,77 @@ class ServerController extends Controller
         $server->description = $request->input('description', $server->description);
         $server->save();
 
-        return redirect()->route('settings.server', ['serverId' => $serverId])
+        //        broadcast(new ServerEdited($serverId, $server->name, $server->description, $server->icon));
+
+        return response()->json(['message' => 'Server updated successfully.']);
+    }
+
+    public function showSettings($serverId)
+    {
+        $server = Server::find($serverId);
+        if (! $server) {
+            return response()->json(['message' => 'Server not found.'], 404);
+        }
+
+        return Inertia::render('Settings/Server')->with([
+            'selected_server' => $server,
+            'selected_server.users' => $server->users,
+            'selected_server.roles' => $server->roles,
+        ]);
+    }
+
+    public function destroy(int $serverId)
+    {
+        $server = Server::find($serverId);
+
+        if (! $server) {
+            return response()->json(['message' => 'Server not found.'], 404);
+        }
+
+        $roles = $server->roles->intersect(Auth::user()->roles);
+
+        if ($roles->doesntContain(function (Role $role) {
+            return $role->hasPerms(PermsType::CAN_DELETE_SERVER->value);
+        })) {
+            return response()->json(['message' => 'Forbidden.'], 403);
+        }
+
+        $server->delete();
+
+        return redirect('/home');
+    }
+
+    public function update(Request $request, int $serverId)
+    {
+        $request->validate([
+            'name' => 'required|string|max:50',
+            'description' => 'nullable|string|max:255',
+            'icon' => 'nullable|image|mimes:jpeg,png,jpg,gif,webp|max:2048',
+        ]);
+
+        $server = Server::find($serverId);
+        if (! $server) {
+            return redirect()->back()->withErrors(['message' => 'Server not found.']);
+        }
+
+        $roles = $server->roles->intersect(Auth::user()->roles);
+
+        if ($roles->doesntContain(function (Role $role) {
+            return $role->hasPerms(PermsType::CAN_EDIT_SERVER->value);
+        })) {
+            return response()->json(['message' => 'Forbidden.'], 403);
+        }
+
+        if ($request->file('icon')?->isValid()) {
+            $path = $request->file('icon')->store('uploads', 'public');
+            $server->icon = Storage::url($path);
+        }
+
+        $server->name = $request->input('name', $server->name);
+        $server->description = $request->input('description', $server->description);
+        $server->save();
+
+        return redirect()->route('settings.server', ['id' => $serverId])
             ->with('message', 'Server updated successfully.');
     }
 
@@ -154,8 +215,16 @@ class ServerController extends Controller
     {
         $server = Server::find($serverId);
 
-        if (!$server) {
+        if (! $server) {
             return response()->json(['message' => 'Server not found.'], 404);
+        }
+
+        $roles = $server->roles->intersect(Auth::user()->roles);
+
+        if ($roles->doesntContain(function (Role $role) {
+            return $role->hasPerms(PermsType::CAN_DELETE_SERVER->value);
+        })) {
+            return response()->json(['message' => 'Forbidden.'], 403);
         }
 
         $server->roles->each(function ($role) {
